Added Scrub pop up slider when Scrub button is pressed.
The scrub button when pressed by the user will now create a popup dialog with a seek bar, much like the WPM button will do. The seekbar is not functional at the moment, will be adding in functionality and ability to slide and get data from the text/epub/URL that is open inside of Glance Reader.