Scrub Button now displaying and operating in UI
The scrub button no longer crashes when the slider is moved. It allows the user to select a position in the text by percentage. Currently does not work on the back end but I added in some variables and code that will be necessary. Had trouble pulling the spritzer data into the ScrubDialogFragment.java class to use for calculations and actually moving the position. Will revisit this later.